+++ spec/terminal_spec.rb
@@ -20,6 +20,19 @@ RSpec.describe ImageUtil::Terminal do
       described_class.detect_support(termin, termout).should == %i[tty sixel]
     end
 
+    it 'detects iterm2 support' do
+      allow(described_class).to receive(:query_terminal).and_return(false, false)
+      termout.remove_instance_variable(:@imageutil_support_cache) if termout.instance_variable_defined?(:@imageutil_support_cache)
+      begin
+        orig = ENV['TERM_PROGRAM']
+        ENV['TERM_PROGRAM'] = 'iTerm.app'
+        described_class.detect_support(termin, termout).should include(:iterm2)
+      ensure
+        ENV['TERM_PROGRAM'] = orig
+        termout.remove_instance_variable(:@imageutil_support_cache) if termout.instance_variable_defined?(:@imageutil_support_cache)
+      end
+    end
+
     it 'caches support results' do
       allow(described_class).to receive(:query_terminal).and_return(false)
       described_class.detect_support(termin, termout).should == %i[tty]
@@ -34,6 +47,13 @@ RSpec.describe ImageUtil::Terminal do
       allow(described_class).to receive(:detect_support).and_return(%i[tty kitty])
       img.should_receive(:to_string).with(:kitty).and_return('K')
       described_class.output_image(termin, termout, img).should == 'K'
+    end
+
+    it 'uses iterm2 when kitty is not available' do
+      img = ImageUtil::Image.new(1, 1)
+      allow(described_class).to receive(:detect_support).and_return(%i[tty iterm2])
+      img.should_receive(:to_string).with(:iterm2).and_return('I')
+      described_class.output_image(termin, termout, img).should == 'I'
     end
 
     it 'uses sixel when kitty is not available' do